refactor memory interface
