+++ pyrit/orchestrator/models/orchestrator_result.py
@@ -92,7 +92,7 @@ class OrchestratorResult:
 
                 if include_auxiliary_scores:
                     auxiliary_scores = (
-                        self._memory.get_scores_by_prompt_ids(prompt_request_response_ids=[str(piece.id)]) or []
+                        self._memory.get_prompt_scores(prompt_ids=[str(piece.id)]) or []
                     )
                     for auxiliary_score in auxiliary_scores:
                         if not self.objective_score or auxiliary_score.id != self.objective_score.id:
