+++ pyrit/orchestrator/multi_turn/multi_turn_orchestrator.py
@@ -167,8 +167,8 @@ class MultiTurnOrchestrator(Orchestrator):
             self._last_prepended_user_message = last_message.converted_value
         elif last_message.role == "assistant":
             # Get scores for the last assistant message based off of the original id
-            self._last_prepended_assistant_message_scores = self._memory.get_scores_by_prompt_ids(
-                prompt_request_response_ids=[str(last_message.original_prompt_id)]
+            self._last_prepended_assistant_message_scores = self._memory.get_prompt_scores(
+                prompt_ids=[str(last_message.original_prompt_id)]
             )
 
             # Do not set last user message if there are no scores for the last assistant message
