+++ tests/unit/memory/memory_interface/test_interface_core.py
@@ -1,0 +1,8 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+from pyrit.memory import MemoryInterface 
+
+
+def test_memory(duckdb_instance: MemoryInterface):
+    assert duckdb_instance
